requirements.txt
django==3.2.25
pymongo==4.6.1
djongo==1.3.6
voyageai==0.2.2
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
sqlparse==0.2.4
django-extensions==3.2.3
